Use logaddexp

+++ src/pigglet/constants.py
@@ -5,6 +5,6 @@ HET_TUP = np.array([False, True], dtype=np.bool_)
 HOM_TUP = np.array([True, False], dtype=np.bool_)
 HET_NUM = 1
 TMP_LABEL = -2
-LOG_LIKE_DTYPE = np.float64
-GL_DTYPE = np.float64
-REAL_SPACE_LIKE_DTYPE = np.float64
+LOG_LIKE_DTYPE = np.float32
+GL_DTYPE = np.float32
+REAL_SPACE_LIKE_DTYPE = np.float32
